verify has member before delete

+++ src/members/members.service.ts
@@ -1,4 +1,8 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { MemberRepository } from './members.repository';
 import { CreateMemberDto } from './dto/create-member.dto';
 import { Member } from './member.model';
@@ -57,6 +61,12 @@ export class MembersService {
   }
 
   async deleteMember(id: string): Promise<void> {
+    const hasMember = await this.memberRepository.hasMember(id);
+
+    if (!hasMember) {
+      throw new NotFoundException(`Member with ID "${id}" not found`);
+    }
+
     return await this.memberRepository.deleteMember(id);
   }
 }
